+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.9.1] - 2024-03-23
 ### Added
 - `from_attribute` and `from_attribue_partial` to `FromAttr`.
 - Support for tuple structs.
@@ -75,7 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking Change**: Compile time verify if ident is given through helper
   trait
 
-[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/ModProg/attribute-derive/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/ModProg/attribute-derive/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/ModProg/attribute-derive/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/ModProg/attribute-derive/compare/v0.7.1...v0.8.0